Extract upload method to dedicated service
